split out functions to separate file
