+++ msp-service/src/index.js
@@ -1,30 +1,21 @@
 import https from "https";
 import http from "http";
-import util from "util";
-import path from "path";
-import fs from "fs";
-import colors from "colors";
+// import util from "util";
+// import path from "path";
+// import fs from "fs";
+// import colors from "colors";
 import winston from "winston";
 import jwt from "jsonwebtoken";
-import url from "url";
+// import url from "url";
 import stringify from "json-stringify-safe";
 import express from "express";
-import moment from "moment";
+// import moment from "moment";
 import {createProxyMiddleware} from "http-proxy-middleware";
 
-// verbose replacement
-function logProvider(provider) {
-    var logger = winston;
+import {logProvider, denyAccess, logSplunkError, logSplunkInfo} from "./helper.js"
 
-    var myCustomProvider = {
-        log: logger.log,
-        debug: logger.debug,
-        info: logSplunkInfo,
-        warn: logger.warn,
-        error: logSplunkError
-    }
-    return myCustomProvider;
-}
+// verbose replacement
+
 
 // winston.add(winston.transports.Console, {
 //    timestamp: true
@@ -281,110 +272,6 @@ app.use('/', proxy);
 // Start express
 app.listen(SERVICE_PORT);
 
-
-/**
- * General deny access handler
- * @param message
- * @param res
- * @param req
- */
-function denyAccess(message, res, req) {
-
-    logSplunkError(message + " - access denied: url: " + stringify(req.originalUrl) + "  request: " + stringify(req.headers));
-
-    res.writeHead(401);
-    res.end();
-}
-
-function logSplunkError (message) {
-
-    // log locally
-    winston.error(message);
-
-    var body = JSON.stringify({
-        message: message
-    })
-
-
-    var options = {
-        hostname: process.env.LOGGER_HOST,
-        port: process.env.LOGGER_PORT,
-        path: '/log',
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json',
-            'Authorization': 'Api-Token ' + process.env.SPLUNK_AUTH_TOKEN,
-            'Content-Length': Buffer.byteLength(body),
-            'logsource': process.env.HOSTNAME,
-            'timestamp': moment().format('DD-MMM-YYYY'),
-            'program': 'msp-service',
-            'severity': 'error'
-        }
-    };
-
-    var req = http.request(options, function (res) {
-        res.setEncoding('utf8');
-        res.on('data', function (chunk) {
-            console.log("Body chunk: " + JSON.stringify(chunk));
-        });
-        res.on('end', function () {
-            console.log('End of chunks');
-        });
-    });
-
-    req.on('error', function (e) {
-        console.error("error sending to splunk-forwarder: " + e.message);
-    });
-
-    // write data to request body
-    req.write(body);
-    req.end();
-}
-
-function logSplunkInfo (message) {
-
-    // log locally
-    winston.info(message);
-
-    var body = JSON.stringify({
-        message: message
-    })
-
-    var options = {
-        hostname: process.env.LOGGER_HOST,
-        port: process.env.LOGGER_PORT,
-        path: '/log',
-        method: 'POST',
-        headers: {
-            'Content-Type': 'application/json',
-            'Authorization': 'Splunk ' + process.env.SPLUNK_AUTH_TOKEN,
-            'Content-Length': Buffer.byteLength(body),
-            'logsource': process.env.HOSTNAME,
-            'timestamp': moment().format('DD-MMM-YYYY'),
-            'method': 'MSP-Service - Pass Through',
-            'program': 'msp-service',
-            'severity': 'info'
-        }
-    };
-
-    var req = http.request(options, function (res) {
-        res.setEncoding('utf8');
-        res.on('data', function (chunk) {
-            console.log("Body chunk: " + JSON.stringify(chunk));
-        });
-        res.on('end', function () {
-            console.log('End of chunks');
-        });
-    });
-
-    req.on('error', function (e) {
-        console.error("error sending to splunk-forwarder: " + e.message);
-    });
-
-    // write data to request body
-    req.write(body);
-    req.end();
-}
 
 logSplunkInfo(`msp-service server started on port ${SERVICE_PORT}`);
